[core] Add dry_run parameters to PurgeFilesProcedure and display the list of deleted directories.

+++ paimon-flink/paimon-flink-1.18/src/main/java/org/apache/paimon/flink/procedure/PurgeFilesProcedure.java
@@ -28,6 +28,7 @@ import org.apache.paimon.table.Table;
 import org.apache.flink.table.procedure.ProcedureContext;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 
 /** A procedure to purge files for a table. */
@@ -36,17 +37,26 @@ public class PurgeFilesProcedure extends ProcedureBase {
 
     public String[] call(ProcedureContext procedureContext, String tableId)
             throws Catalog.TableNotExistException {
+        return call(procedureContext, tableId, false);
+    }
+
+    public String[] call(ProcedureContext procedureContext, String tableId, boolean dryRun)
+            throws Catalog.TableNotExistException {
         Table table = catalog.getTable(Identifier.fromString(tableId));
         FileStoreTable fileStoreTable = (FileStoreTable) table;
         FileIO fileIO = fileStoreTable.fileIO();
         Path tablePath = fileStoreTable.snapshotManager().tablePath();
+        ArrayList<String> deleteDir = new ArrayList<>();
         try {
             Arrays.stream(fileIO.listStatus(tablePath))
                     .filter(f -> !f.getPath().getName().contains("schema"))
                     .forEach(
                             fileStatus -> {
                                 try {
-                                    fileIO.delete(fileStatus.getPath(), true);
+                                    deleteDir.add(fileStatus.getPath().getName());
+                                    if (!dryRun) {
+                                        fileIO.delete(fileStatus.getPath(), true);
+                                    }
                                 } catch (IOException e) {
                                     throw new RuntimeException(e);
                                 }
@@ -54,9 +64,9 @@ public class PurgeFilesProcedure extends ProcedureBase {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        return new String[] {
-            String.format("Success purge files with table: %s.", fileStoreTable.name())
-        };
+        return deleteDir.isEmpty()
+                ? new String[] {"There are no dir to be deleted."}
+                : deleteDir.toArray(new String[0]);
     }
 
     @Override
